+++ base32ct/src/encoding.rs
@@ -220,14 +220,9 @@ impl<T: Alphabet> Encoding for T {
         }
     }
 
+    #[inline]
     fn encoded_len(bytes: &[u8]) -> usize {
-        if bytes.is_empty() {
-            0
-        } else if Self::PADDED {
-            ((bytes.len() - 1) / 5 + 1) * 8
-        } else {
-            (bytes.len() * 8) / 5 + 1
-        }
+        encoded_len::<Self>(bytes.len())
     }
 }
 
@@ -257,4 +252,86 @@ fn remove_padding(mut input: &[u8]) -> Result<&[u8]> {
     }
 
     Ok(input)
+}
+
+/// Get the length of Base32 produced by encoding the given amount of bytes.
+pub const fn encoded_len<T: Encoding>(length: usize) -> usize {
+    if length == 0 {
+        0
+    } else if T::PADDED {
+        ((length - 1) / 5 + 1) * 8
+    } else {
+        (length * 8 + 4) / 5
+    }
+}
+
+#[cfg(all(test, feature = "alloc"))]
+mod tests {
+    use crate::{Base32, Base32Unpadded, Encoding};
+
+    struct LenData {
+        fourty_bit_groups_len: usize,
+        last_group_len: usize,
+        padding_len: usize,
+    }
+
+    fn get_len_data(data_len: usize) -> LenData {
+        // More information about the calculation can be found at
+        // https://www.rfc-editor.org/rfc/rfc4648#section-6
+        let fourty_bit_groups_len = data_len / 5 * 8;
+        let (last_group_len, padding_len) = match data_len % 5 {
+            0 => (0, 0),
+            1 => (2, 6),
+            2 => (4, 4),
+            3 => (5, 3),
+            4 => (7, 1),
+            _ => unreachable!(),
+        };
+
+        LenData {
+            fourty_bit_groups_len,
+            last_group_len,
+            padding_len,
+        }
+    }
+
+    #[test]
+    fn unpadded_encoded_len() {
+        let mut buf = vec![];
+        assert_eq!(Base32Unpadded::encoded_len(&buf), 0);
+
+        for _ in 0..10 {
+            buf.push(b'a');
+            let LenData {
+                fourty_bit_groups_len,
+                last_group_len,
+                padding_len: _,
+            } = get_len_data(buf.len());
+
+            assert_eq!(
+                Base32Unpadded::encoded_len(&buf),
+                fourty_bit_groups_len + last_group_len
+            );
+        }
+    }
+
+    #[test]
+    fn padded_encoded_len() {
+        let mut buf = vec![];
+        assert_eq!(Base32::encoded_len(&buf), 0);
+
+        for _ in 0..10 {
+            buf.push(b'a');
+            let LenData {
+                fourty_bit_groups_len,
+                last_group_len,
+                padding_len,
+            } = get_len_data(buf.len());
+
+            assert_eq!(
+                Base32::encoded_len(&buf),
+                fourty_bit_groups_len + last_group_len + padding_len,
+            );
+        }
+    }
 }