+++ base32ct/src/lib.rs
@@ -58,7 +58,7 @@ mod encoding;
 mod error;
 
 pub use crate::{
-    alphabet::rfc4648::{Base32, Base32Unpadded, Base32Upper},
-    encoding::Encoding,
+    alphabet::rfc4648::{Base32, Base32Unpadded, Base32Upper, Base32UpperUnpadded},
+    encoding::{encoded_len, Encoding},
     error::{Error, Result},
 };