+++ base32ct/tests/vectors.rs
@@ -2,7 +2,7 @@
 
 #![cfg(feature = "alloc")]
 
-use base32ct::{Base32, Base32Unpadded, Base32Upper, Encoding, Error};
+use base32ct::{Base32, Base32Unpadded, Base32Upper, Base32UpperUnpadded, Encoding, Error};
 
 #[derive(Debug)]
 struct TestVector {
@@ -23,6 +23,18 @@ const LOWER_PADDED_VECTORS: &[TestVector] = &[
         decoded: &[32, 7],
         encoded: "eadq====",
     },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56],
+        encoded: "ci2fm===",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56, 0x78, 0x9a],
+        encoded: "ci2fm6e2",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56, 0x78, 0x9a, 0xbc],
+        encoded: "ci2fm6e2xq======",
+    },
 ];
 
 const LOWER_UNPADDED_VECTORS: &[TestVector] = &[
@@ -38,6 +50,18 @@ const LOWER_UNPADDED_VECTORS: &[TestVector] = &[
         decoded: &[32, 7],
         encoded: "eadq",
     },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56],
+        encoded: "ci2fm",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56, 0x78, 0x9a],
+        encoded: "ci2fm6e2",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56, 0x78, 0x9a, 0xbc],
+        encoded: "ci2fm6e2xq",
+    },
 ];
 
 const UPPER_PADDED_VECTORS: &[TestVector] = &[
@@ -52,6 +76,45 @@ const UPPER_PADDED_VECTORS: &[TestVector] = &[
     TestVector {
         decoded: &[32, 7],
         encoded: "EADQ====",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56],
+        encoded: "CI2FM===",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56, 0x78, 0x9a],
+        encoded: "CI2FM6E2",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56, 0x78, 0x9a, 0xbc],
+        encoded: "CI2FM6E2XQ======",
+    },
+];
+
+const UPPER_UNPADDED_VECTORS: &[TestVector] = &[
+    TestVector {
+        decoded: &[0],
+        encoded: "AA",
+    },
+    TestVector {
+        decoded: &[1, 2, 3, 5, 9, 17, 33, 65, 129],
+        encoded: "AEBAGBIJCEQUDAI",
+    },
+    TestVector {
+        decoded: &[32, 7],
+        encoded: "EADQ",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56],
+        encoded: "CI2FM",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56, 0x78, 0x9a],
+        encoded: "CI2FM6E2",
+    },
+    TestVector {
+        decoded: &[0x12, 0x34, 0x56, 0x78, 0x9a, 0xbc],
+        encoded: "CI2FM6E2XQ",
     },
 ];
 
@@ -71,6 +134,13 @@ fn decode_valid_base32() {
     for vector in UPPER_PADDED_VECTORS {
         assert_eq!(
             &Base32Upper::decode_vec(vector.encoded).unwrap(),
+            vector.decoded
+        );
+    }
+
+    for vector in UPPER_UNPADDED_VECTORS {
+        assert_eq!(
+            &Base32UpperUnpadded::decode_vec(vector.encoded).unwrap(),
             vector.decoded
         );
     }
@@ -106,5 +176,12 @@ fn encode_base32() {
 
     for vector in UPPER_PADDED_VECTORS {
         assert_eq!(&Base32Upper::encode_string(vector.decoded), vector.encoded);
+    }
+
+    for vector in UPPER_UNPADDED_VECTORS {
+        assert_eq!(
+            &Base32UpperUnpadded::encode_string(vector.decoded),
+            vector.encoded
+        );
     }
 }