+++ base32ct/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.1 (2024-05-28)
+### Added
+- Support for Base32 upper unpadded alphabet ([#1406])
+
+### Fixed
+- Broken encoding of unpadded base32 ([#1421])
+
+[#1406]: https://github.com/RustCrypto/formats/pull/1406
+[#1421]: https://github.com/RustCrypto/formats/pull/1421
+
 ## 0.2.0 (2023-02-26)
 ### Changed
 - MSRV 1.60 ([#802])